requirements.txt
asgiref==3.3.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
install==1.3.4
pytz==2020.4
sqlparse==0.4.1
